Split relationship matching into helper module
