+++ lib/rspec-puppet.rb
@@ -4,6 +4,7 @@ require 'rspec/its'
 require 'fileutils'
 require 'tmpdir'
 require 'rspec-puppet/errors'
+require 'rspec-puppet/helpers'
 require 'rspec-puppet/matchers'
 require 'rspec-puppet/example'
 require 'rspec-puppet/setup'
