+++ lib/rspec-puppet/helpers.rb
@@ -1,0 +1,6 @@
+module RSpec::Puppet
+  module Helpers
+  end
+end
+
+require 'rspec-puppet/helpers/relationships'
